pyproject.toml
[tool.poetry]
name = "base-fastapi"
version = "0.1.0"
description = "This is a base project structure for fastapi"
authors = ["gunblasterXTO"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
email-validator = "^2.1.0.post1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
ruff = "^0.1.11"
mypy = "^1.8.0"
alembic = "^1.13.1"
coverage = "^7.4.1"
bandit = "^1.7.7"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"
pre-commit = "^3.6.0"

[tool.ruff]
line-length = 79
src = ["app", "tests"]
preview = true

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.format]
docstring-code-line-length = 60
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true
case-sensitive = true
detect-same-package = true
force-wrap-aliases = true
order-by-type = true
known-first-party = ["app"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
